carousel panels central
solved the problem with the carousel panels not being central. it seems that the grid previously used was causing the problem. although with mobile screen resolution still doesn't behave as expected. to solve - contain different widths diferent sections -  function carousel (at the moment the first carousel is the only to fuction as it should the others when clicked make the first to respond)