starting to work on the different carousels
starting to work on the different carousels. At the moment because they share the same classes and id the first one has priority over the others indicators, this is when pressed the indicator of b or c carousel only the a responds. Also changing the indicators and controls for the ones in the design add the new controlers